Moved a few files relating to Firebase services into service directory. Renamed best_total_score in stat_prefs.proto to best_combined_score to better represent its use. Created SingleGameStats which will be used to store/retrieve data from Firestore.
