+++ app/src/main/java/com/heyzeusv/solitaire/menu/settings/SettingsMenu.kt
@@ -61,6 +61,8 @@ import com.heyzeusv.solitaire.R
 import com.heyzeusv.solitaire.menu.MenuScreen
 import com.heyzeusv.solitaire.menu.MenuViewModel
 import com.heyzeusv.solitaire.board.animation.AnimationDurations
+import com.heyzeusv.solitaire.service.AccountStatus
+import com.heyzeusv.solitaire.service.UserData
 import com.heyzeusv.solitaire.util.AccountTabs
 import com.heyzeusv.solitaire.util.MenuState
 import com.heyzeusv.solitaire.util.PreviewUtil
