+++ app/src/main/java/com/heyzeusv/solitaire/service/SingleGameStats.kt
@@ -1,0 +1,13 @@
+package com.heyzeusv.solitaire.service
+
+data class SingleGameStats(
+    val game: String,
+    val gamesPlayed: Int,
+    val gamesWon: Int,
+    val lowestMoves: Int,
+    val totalMoves: Int,
+    val fastestWin: Long,
+    val totalTime: Long,
+    val totalScore: Int,
+    val bestCombinedScore: Long
+)
